requirements.txt
kivymd
mysql-connector-python
opencv-python 
face_recognition 
numpy